(refactor): extracted restrictions by field to separate class.
