+++ amf-shapes/shared/src/main/scala/amf/shapes/internal/domain/resolution/shape_normalization/RestrictionComputation.scala
@@ -7,7 +7,7 @@ import amf.core.internal.metamodel.domain.extensions.PropertyShapeModel
 import amf.core.internal.metamodel.{Field, Obj}
 import amf.core.internal.parser.domain.Annotations
 import amf.shapes.internal.domain.metamodel._
-import amf.shapes.internal.domain.resolution.shape_normalization.AmfElementComparer._
+import amf.shapes.internal.domain.resolution.shape_normalization.restrictions._
 
 private[shape_normalization] trait RestrictionComputation {
 
@@ -68,454 +68,28 @@ private[shape_normalization] trait RestrictionComputation {
     shape
   }
 
-  protected def computeNumericRestriction(
-      comparison: String,
-      lvalue: AmfElement,
-      rvalue: AmfElement
-  ): AmfElement = {
-    (lvalue, rvalue) match {
-      case BothNumeric(lnum, rnum) =>
-        comparison match {
-          case "max" =>
-            if (lnum.intValue() <= rnum.intValue()) {
-              rvalue
-            } else {
-              lvalue
-            }
-          case "min" =>
-            if (lnum.intValue() >= rnum.intValue()) {
-              rvalue
-            } else {
-              lvalue
-            }
-          case _ => throw new InheritanceIncompatibleShapeError(s"Unknown numeric comparison $comparison")
-        }
-      case _ =>
-        throw new InheritanceIncompatibleShapeError("Cannot compare non numeric or missing values")
-    }
-  }
-
-  protected def computeEnum(
-      derivedEnumeration: Seq[AmfElement],
-      superEnumeration: Seq[AmfElement],
-      annotations: Annotations
-  ): Unit = {
-    if (derivedEnumeration.nonEmpty && superEnumeration.nonEmpty) {
-      val headOption = derivedEnumeration.headOption
-      if (headOption.exists(h => superEnumeration.headOption.exists(_.getClass != h.getClass)))
-        throw new InheritanceIncompatibleShapeError(
-          s"Values in subtype enumeration are from different class '${derivedEnumeration.head.getClass}' of the super type enumeration '${superEnumeration.head.getClass}'",
-          Some(ShapeModel.Values.value.iri()),
-          headOption.flatMap(_.location()),
-          headOption.flatMap(_.position())
-        )
-
-      derivedEnumeration match {
-        case Seq(_: ScalarNode) =>
-          val superScalars = superEnumeration.collect({ case s: ScalarNode => s.value.value() })
-          val ds           = derivedEnumeration.asInstanceOf[Seq[ScalarNode]]
-          ds.foreach { e =>
-            if (!superScalars.contains(e.value.value())) {
-              throw new InheritanceIncompatibleShapeError(
-                s"Values in subtype enumeration (${ds.map(_.value).mkString(",")}) not found in the supertype enumeration (${superScalars
-                    .mkString(",")})",
-                Some(ShapeModel.Values.value.iri()),
-                e.location(),
-                e.position()
-              )
-            }
-          }
-        case _ => // ignore
-      }
-    }
-  }
-
-  protected def maybeAsString(value: AmfElement): Option[String] = {
-    value match {
-      case scalar: AmfScalar => Option(scalar.value).map(_.toString)
-      case _                 => None
-    }
-  }
-
   protected def computeNarrow(field: Field, derivedValue: AmfElement, superValue: AmfElement): AmfElement = {
     field match {
-
-      case ShapeModel.Name =>
-        val derivedStrValue = maybeAsString(derivedValue)
-        val superStrValue   = maybeAsString(superValue)
-        (superStrValue, derivedStrValue) match {
-          case (Some(_), None | Some("schema")) => superValue
-          case _                                => derivedValue
-        }
-
-      case NodeShapeModel.MinProperties =>
-        if (
-          lessOrEqualThan(
-            superValue,
-            derivedValue,
-            incompatibleException(NodeShapeModel.MinProperties, derivedValue)
-          )
-        ) {
-          computeNumericRestriction(
-            "max",
-            superValue,
-            derivedValue
-          )
-        } else {
-          throw new InheritanceIncompatibleShapeError(
-            "Resolution error: sub type has a weaker constraint for min-properties than base type for minProperties",
-            Some(NodeShapeModel.MinProperties.value.iri()),
-            derivedValue.location(),
-            derivedValue.position()
-          )
-        }
-
-      case NodeShapeModel.MaxProperties =>
-        if (
-          moreOrEqualThan(
-            superValue,
-            derivedValue,
-            incompatibleException(NodeShapeModel.MaxProperties, derivedValue)
-          )
-        ) {
-          computeNumericRestriction(
-            "min",
-            superValue,
-            derivedValue
-          )
-        } else {
-          throw new InheritanceIncompatibleShapeError(
-            "Resolution error: sub type has a weaker constraint for max-properties than base type for maxProperties",
-            Some(NodeShapeModel.MaxProperties.value.iri()),
-            derivedValue.location(),
-            derivedValue.position()
-          )
-        }
-
-      case ScalarShapeModel.MinLength =>
-        if (
-          lessOrEqualThan(
-            superValue,
-            derivedValue,
-            incompatibleException(ScalarShapeModel.MinLength, derivedValue)
-          )
-        ) {
-          computeNumericRestriction(
-            "max",
-            superValue,
-            derivedValue
-          )
-        } else {
-          throw new InheritanceIncompatibleShapeError(
-            "Resolution error: sub type has a weaker constraint for min-length than base type for maxProperties",
-            Some(ScalarShapeModel.MinLength.value.iri()),
-            derivedValue.location(),
-            derivedValue.position()
-          )
-        }
-
-      case ScalarShapeModel.MaxLength =>
-        if (
-          moreOrEqualThan(
-            superValue,
-            derivedValue,
-            incompatibleException(ScalarShapeModel.MaxLength, derivedValue)
-          )
-        ) {
-          computeNumericRestriction(
-            "min",
-            superValue,
-            derivedValue
-          )
-        } else {
-          throw new InheritanceIncompatibleShapeError(
-            "Resolution error: sub type has a weaker constraint for max-length than base type for maxProperties",
-            Some(ScalarShapeModel.MaxLength.value.iri()),
-            derivedValue.location(),
-            derivedValue.position()
-          )
-        }
-
-      case ScalarShapeModel.Minimum =>
-        if (
-          lessOrEqualThan(
-            superValue,
-            derivedValue,
-            incompatibleException(ScalarShapeModel.Minimum, derivedValue)
-          )
-        ) {
-          computeNumericRestriction(
-            "max",
-            superValue,
-            derivedValue
-          )
-        } else {
-          throw new InheritanceIncompatibleShapeError(
-            "Resolution error: sub type has a weaker constraint for min-minimum than base type for minimum",
-            Some(ScalarShapeModel.Minimum.value.iri()),
-            derivedValue.location(),
-            derivedValue.position()
-          )
-        }
-
-      case ScalarShapeModel.Maximum =>
-        if (
-          moreOrEqualThan(
-            superValue,
-            derivedValue,
-            incompatibleException(ScalarShapeModel.Maximum, derivedValue)
-          )
-        ) {
-          computeNumericRestriction(
-            "min",
-            superValue,
-            derivedValue
-          )
-        } else {
-          throw new InheritanceIncompatibleShapeError(
-            "Resolution error: sub type has a weaker constraint for maximum than base type for maximum",
-            Some(ScalarShapeModel.Maximum.value.iri()),
-            derivedValue.location(),
-            derivedValue.position()
-          )
-        }
-
-      case ArrayShapeModel.MinItems =>
-        if (
-          lessOrEqualThan(
-            superValue,
-            derivedValue,
-            incompatibleException(ArrayShapeModel.MinItems, derivedValue)
-          )
-        ) {
-          computeNumericRestriction(
-            "max",
-            superValue,
-            derivedValue
-          )
-        } else {
-          throw new InheritanceIncompatibleShapeError(
-            "Resolution error: sub type has a weaker constraint for minItems than base type for minItems",
-            Some(ArrayShapeModel.MinItems.value.iri()),
-            derivedValue.location(),
-            derivedValue.position(),
-            isViolation = true
-          )
-        }
-
-      case ArrayShapeModel.MaxItems =>
-        if (
-          moreOrEqualThan(
-            superValue,
-            derivedValue,
-            incompatibleException(ArrayShapeModel.MaxItems, derivedValue)
-          )
-        ) {
-          computeNumericRestriction(
-            "min",
-            superValue,
-            derivedValue
-          )
-        } else {
-          throw new InheritanceIncompatibleShapeError(
-            "Resolution error: sub type has a weaker constraint for maxItems than base type for maxItems",
-            Some(ArrayShapeModel.MaxItems.value.iri()),
-            derivedValue.location(),
-            derivedValue.position()
-          )
-        }
-
-      case ScalarShapeModel.Format =>
-        if (
-          areEqualStrings(
-            superValue,
-            derivedValue,
-            incompatibleException(ScalarShapeModel.Format, derivedValue)
-          )
-        ) {
-          derivedValue
-        } else {
-          throw new InheritanceIncompatibleShapeError(
-            "different values for format constraint",
-            Some(ScalarShapeModel.Format.value.iri()),
-            derivedValue.location(),
-            derivedValue.position()
-          )
-        }
-
-      case ScalarShapeModel.Pattern =>
-        if (
-          areEqualStrings(
-            superValue,
-            derivedValue,
-            incompatibleException(ScalarShapeModel.Pattern, derivedValue)
-          )
-        ) {
-          derivedValue
-        } else {
-          throw new InheritanceIncompatibleShapeError(
-            "different values for pattern constraint",
-            Some(ScalarShapeModel.Pattern.value.iri()),
-            derivedValue.location(),
-            derivedValue.position()
-          )
-        }
-
-      case NodeShapeModel.Discriminator =>
-        if (
-          !areEqualStrings(
-            superValue,
-            derivedValue,
-            incompatibleException(NodeShapeModel.Discriminator, derivedValue)
-          )
-        ) {
-          derivedValue
-        } else {
-          throw new InheritanceIncompatibleShapeError(
-            "shape has same discriminator value as parent",
-            Some(NodeShapeModel.Discriminator.value.iri()),
-            derivedValue.location(),
-            derivedValue.position()
-          )
-        }
-
-      case NodeShapeModel.DiscriminatorValue =>
-        if (
-          !areEqualStrings(
-            superValue,
-            derivedValue,
-            incompatibleException(NodeShapeModel.DiscriminatorValue, derivedValue)
-          )
-        ) {
-          derivedValue
-        } else {
-          throw new InheritanceIncompatibleShapeError(
-            "shape has same discriminator value as parent",
-            Some(NodeShapeModel.DiscriminatorValue.value.iri()),
-            derivedValue.location(),
-            derivedValue.position()
-          )
-        }
-
-      case ShapeModel.Values =>
-        computeEnum(
-          derivedValue.asInstanceOf[AmfArray].values,
-          superValue.asInstanceOf[AmfArray].values,
-          derivedValue.annotations
-        )
-        derivedValue
-
-      case ArrayShapeModel.UniqueItems =>
-        if (
-          areEqualBooleans(
-            superValue,
-            derivedValue,
-            incompatibleException(ArrayShapeModel.UniqueItems, derivedValue)
-          ) ||
-          areEqualBooleans(
-            superValue,
-            derivedValue,
-            incompatibleException(ArrayShapeModel.UniqueItems, derivedValue)
-          ) ||
-          areExpectedBooleans(
-            superValue,
-            derivedValue,
-            expectedLeft = false,
-            expectedRight = true,
-            incompatibleException(ArrayShapeModel.UniqueItems, derivedValue)
-          )
-        ) {
-          derivedValue
-        } else {
-          throw new InheritanceIncompatibleShapeError(
-            "different values for unique items constraint",
-            Some(ArrayShapeModel.UniqueItems.value.iri()),
-            derivedValue.location(),
-            derivedValue.position()
-          )
-        }
-
-      case PropertyShapeModel.MinCount =>
-        if (
-          lessOrEqualThan(
-            superValue,
-            derivedValue,
-            incompatibleException(PropertyShapeModel.MinCount, derivedValue)
-          )
-        ) {
-          computeNumericRestriction(
-            "max",
-            superValue,
-            derivedValue
-          )
-        } else {
-          throw new InheritanceIncompatibleShapeError(
-            "Resolution error: sub type has a weaker constraint for minCount than base type for minCount",
-            Some(PropertyShapeModel.MinCount.value.iri()),
-            derivedValue.location(),
-            derivedValue.position()
-          )
-        }
-
-      case PropertyShapeModel.MaxCount =>
-        if (
-          moreOrEqualThan(
-            superValue,
-            derivedValue,
-            incompatibleException(PropertyShapeModel.MaxCount, derivedValue)
-          )
-        ) {
-          computeNumericRestriction(
-            "min",
-            superValue,
-            derivedValue
-          )
-        } else {
-          throw new InheritanceIncompatibleShapeError(
-            "Resolution error: sub type has a weaker constraint for maxCount than base type for maxCount",
-            Some(PropertyShapeModel.MaxCount.value.iri()),
-            derivedValue.location(),
-            derivedValue.position()
-          )
-        }
-
-      case PropertyShapeModel.Path =>
-        if (
-          areEqualStrings(
-            superValue,
-            derivedValue,
-            incompatibleException(PropertyShapeModel.Path, derivedValue)
-          )
-        ) {
-          derivedValue
-        } else {
-          throw new InheritanceIncompatibleShapeError(
-            "different values for discriminator value path",
-            Some(PropertyShapeModel.Path.value.iri()),
-            derivedValue.location(),
-            derivedValue.position()
-          )
-        }
-
-      case PropertyShapeModel.Range =>
-        if (
-          areEqualStrings(
-            superValue,
-            derivedValue,
-            incompatibleException(PropertyShapeModel.Range, derivedValue)
-          )
-        ) {
-          derivedValue
-        } else {
-          throw new InheritanceIncompatibleShapeError(
-            "different values for discriminator value range",
-            Some(PropertyShapeModel.Range.value.iri()),
-            derivedValue.location(),
-            derivedValue.position()
-          )
-        }
-
-      case _ => derivedValue
+      case ShapeModel.Name                   => NameRestriction.restrict(field, derivedValue, superValue)
+      case NodeShapeModel.MinProperties      => MinPropertiesRestriction.restrict(field, derivedValue, superValue)
+      case NodeShapeModel.MaxProperties      => MaxPropertiesRestriction.restrict(field, derivedValue, superValue)
+      case ScalarShapeModel.MinLength        => MinLengthRestriction.restrict(field, derivedValue, superValue)
+      case ScalarShapeModel.MaxLength        => MaxLengthRestriction.restrict(field, derivedValue, superValue)
+      case ScalarShapeModel.Minimum          => MinimumRestriction.restrict(field, derivedValue, superValue)
+      case ScalarShapeModel.Maximum          => MaximumRestriction.restrict(field, derivedValue, superValue)
+      case ArrayShapeModel.MinItems          => MinItemsRestriction.restrict(field, derivedValue, superValue)
+      case ArrayShapeModel.MaxItems          => MaxItemsRestriction.restrict(field, derivedValue, superValue)
+      case ScalarShapeModel.Format           => FormatRestriction.restrict(field, derivedValue, superValue)
+      case ScalarShapeModel.Pattern          => PatternRestriction.restrict(field, derivedValue, superValue)
+      case NodeShapeModel.Discriminator      => DiscriminatorRestriction.restrict(field, derivedValue, superValue)
+      case NodeShapeModel.DiscriminatorValue => DiscriminatorValueRestriction.restrict(field, derivedValue, superValue)
+      case ShapeModel.Values                 => ValuesRestriction.restrict(field, derivedValue, superValue)
+      case ArrayShapeModel.UniqueItems       => UniqueItemsRestriction.restrict(field, derivedValue, superValue)
+      case PropertyShapeModel.MinCount       => MinCountRestriction.restrict(field, derivedValue, superValue)
+      case PropertyShapeModel.MaxCount       => MaxCountRestriction.restrict(field, derivedValue, superValue)
+      case PropertyShapeModel.Path           => PathRestriction.restrict(field, derivedValue, superValue)
+      case PropertyShapeModel.Range          => RangeRestriction.restrict(field, derivedValue, superValue)
+      case _                                 => derivedValue
     }
   }
 }
